+++ demo07/src/it/unito/jas2/demo07/model/PersonsModel.java
@@ -41,10 +41,10 @@ public class PersonsModel extends AbstractSimulationManager implements EventList
 	private static Logger log = Logger.getLogger(PersonsModel.class);
 	
 	@ModelParameter(description="Simulation first year (valid range 2002-2060)")
-	private Integer startYear = 2002;
+	private Double startYear = 2002.0;
 
 	@ModelParameter(description="Simulation ends at year [valid range 2003-2061]")
-	private Integer endYear = 2061;
+	private Double endYear = 2061.0;
 	
 	@ModelParameter(description="Retirement age for women")
 	private Integer wemra = 61;
@@ -193,7 +193,7 @@ public class PersonsModel extends AbstractSimulationManager implements EventList
 		//Schedule model to stop
 		getEngine().getEventList().scheduleOnce(new SingleTargetEvent(this, Processes.Stop), endYear, Order.AFTER_ALL.getOrdering());
 		
-		year = startYear;
+		year = startYear.intValue();
 		elapsedTime = System.currentTimeMillis();
 		
 	}
@@ -295,7 +295,7 @@ public class PersonsModel extends AbstractSimulationManager implements EventList
 			MultiKey mk = (MultiKey) iterator.getKey();
 			Integer ageFrom = (Integer) mk.getKey(0);
 			Integer ageTo = (Integer) mk.getKey(1);
-			double divorceTarget = ((Number) pDivorceMap.getValue(ageFrom, ageTo, SimulationEngine.getInstance().getTime())).doubleValue();
+			double divorceTarget = ((Number) pDivorceMap.getValue(ageFrom, ageTo, (int)SimulationEngine.getInstance().getTime())).doubleValue();
 
 			//Align
 			new SBDAlignment<Person>().align(
@@ -331,7 +331,7 @@ public class PersonsModel extends AbstractSimulationManager implements EventList
 //			Gender gender = Gender.values()[(Integer) mk.getKey(2)];		//When 1 / 0 are entries for Male / Female respectively in p_inwork.xls
 			Gender gender = Gender.valueOf((String) mk.getKey(2));			//When Male / Female are entries in p_inwork.xls
 
-			double inWorkTarget = ((Number) map.getValue(ageFrom, ageTo, gender.toString(), SimulationEngine.getInstance().getTime())).doubleValue();
+			double inWorkTarget = ((Number) map.getValue(ageFrom, ageTo, gender.toString(), (int)SimulationEngine.getInstance().getTime())).doubleValue();
 
 			//Align
 			new SBDAlignment<Person>().align(
@@ -409,7 +409,7 @@ public class PersonsModel extends AbstractSimulationManager implements EventList
 	// Access methods
 	// ---------------------------------------------------------------------
 	
-	public Integer getStartYear() {
+	public Double getStartYear() {
 		return startYear;
 	}
 
@@ -421,7 +421,7 @@ public class PersonsModel extends AbstractSimulationManager implements EventList
 		return households;
 	}
 	
-	public Integer getEndYear() {
+	public Double getEndYear() {
 		return endYear;
 	}
